token blocks processing into tokens.rs
